Implemented the shop details screen using id and fixed the saved shop toggl
